+++ dashboard/src/views/GrammateiaView.vue
@@ -19,7 +19,7 @@
           <td class="text-xs-left">{{ props.item.arithmosMitroou }}</td>
           <td class="text-xs-left">{{ props.item.semester }}</td>
           <td style="padding:0 0 0 0 ">
-            <v-btn round small @click="deleteStudent(props.item.arithmosMitroou)">Διαγραφη</v-btn>
+            <v-btn round small @click="sd(props.item.arithmosMitroou)">Διαγραφη</v-btn>
           </td>
         </template>
         <v-alert
@@ -30,19 +30,55 @@
         >Your search for "{{ search }}" found no results.</v-alert>
       </v-data-table>
     </v-card>
-            <v-btn round @click="semesterBegin()">Εκκινηση νεου εξαμηνου</v-btn>
+    <v-btn @click="dialog = true" round>Εκκινηση νεου εξαμηνου</v-btn>
+    <v-dialog v-model="dialog" persistent max-width="290">
+      <v-card>
+        <v-card-title class="headline">Εκκινηση νεου εξαμηνου</v-card-title>
+        <v-card-text>ΠΡΟΣΟΧΗ! Αν επιβεβαιωσετε το εξαμηνο θα ανεβει κατα ενα και θα επηρεασει ολα τα δεδομενα. Ειστε σιγουροι για την ανανεωση του εξαμηνου</v-card-text>
+        <v-card-actions>
+          <v-spacer></v-spacer>
+          <v-btn color="pink darken-1" flat @click="dialog = false">Disagree</v-btn>
+          <v-btn color="pink darken-1" flat @click="semesterBegin()">Agree</v-btn>
+        </v-card-actions>
+      </v-card>
+    </v-dialog>
+    <v-dialog v-model="dialog1" persistent max-width="290">
+      <v-card>
+        <v-card-title class="headline">Διαγραφη χρηστη</v-card-title>
+        <v-card-text>ΠΡΟΣΟΧΗ! Ειστε σιγουροι οτι θελετε να διαγραψετε τον συγκεκριμενο χρηστη;</v-card-text>
+        <v-card-actions>
+          <v-spacer></v-spacer>
+          <v-btn color="pink darken-1" flat @click="dialog1 = false">Disagree</v-btn>
+          <v-btn color="pink darken-1" flat @click="deleteStudent(studentNumber)">Agree</v-btn>
+        </v-card-actions>
+      </v-card>
+    </v-dialog>
+    <v-dialog v-model="dialog3" max-width="290">
+      <v-card>
+        <v-card-title class="headline">Παρακαλω εισαγετε δακτυλικο αποτυπωμα</v-card-title>
+        <v-card-text class="text-xs-center"><v-icon x-large>{{fingericon}}</v-icon></v-card-text>
+      </v-card>
+    </v-dialog>
   </v-flex>
 </template>
 
 <script>
 import StudentService from '@/services/StudentService'
 import UserService from '@/services/UserService'
+import FingerprintService from '@/services/FingerprintService'
+import util from '@/utils/util.js'
+
 export default {
   name: 'grammateia',
   data () {
     return {
       dialog: false,
+      dialog1: false,
+      dialog3: false,
+      response1: '',
+      fingericon: 'fas fa-fingerprint',
       loader: null,
+      studentNumber: 0,
       loading: false,
       loading1: false,
       search: '',
@@ -60,7 +96,6 @@ export default {
         // { text: 'Περασμένα Μαθήματα', value: 'perasmena' }
       ],
       students: []
-
     }
   },
   watch: {},
@@ -71,25 +106,57 @@ export default {
     async getAllStudents () {
       try {
         this.students = await StudentService.getAllStudents()
-        console.log(this.students)
       } catch (error) {
-        window.alert(error)
+        console.log(error)
       }
-      console.log(this.students)
     },
     async deleteStudent (arithmosMitroou) {
       try {
-        let response = await StudentService.deleteStudent(arithmosMitroou)
-        if (response.deletedCount === 1) {
-          location.reload(true)
+        // eslint-disable-next-line
+        if (this.response1.fingerprint[0].fid == this.$store.getters.user.fingerprintID) {
+          let response = await StudentService.deleteStudent(arithmosMitroou)
+          if (response.deletedCount === 1) {
+            await FingerprintService.cleanFingerprint()
+            location.reload(true)
+          }
+        } else {
+          await FingerprintService.cleanFingerprint()
+          this.dialog1 = false
+          window.alert('not correct fingerprint')
         }
       } catch (error) {
-        window.alert(error)
+        console.log(error)
       }
     },
     async semesterBegin () {
-      let semester = await UserService.semesterBegin()
-      console.log(semester)
+      try {
+        let semester = await UserService.semesterBegin()
+        if (semester.data.success === true) {
+          location.reload(true)
+        }
+      } catch (error) {
+        console.log(error)
+      }
+    },
+    async sd (sn) {
+      try {
+        this.dialog3 = true
+        this.response1 = await FingerprintService.getFingerprint()
+        this.studentNumber = sn
+        while (this.response1.fingerprint.length === 0) {
+          let toConfirm = await FingerprintService.getFingerprint()
+          if (toConfirm.fingerprint !== undefined) {
+            this.response1 = toConfirm
+          }
+          await util.sleep(1500)
+        }
+        if (this.response1.fingerprint.length !== 0) {
+          this.dialog3 = false
+          this.dialog1 = true
+        }
+      } catch (error) {
+        console.log(error)
+      }
     }
   }
 }
